WebDevLab03/requirements.txt
streamlit
google-generativeai
requests
matplotlib.pyplot
pandas
